# openapi.yaml
openapi: 3.0.3
info:
  title: SmartCloudOps AI API
  description: Enterprise AI-powered cloud operations platform with real-time monitoring,
    anomaly detection, and automated remediation
  version: 5.0.0
  contact:
    name: SmartCloudOps Team
    email: support@smartcloudops.ai
    url: https://smartcloudops.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Development server
- url: https://api.smartcloudops.ai
  description: Production server
security:
- bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and receive JWT token
      tags:
      - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password123
      required:
      - username
      - password
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
            role:
              type: string
...